add joke Api
