+++ src/server.js
@@ -5,17 +5,13 @@ const cors = require('cors');
 const morgan = require('morgan');
 const logger = require('./middleware/logger');
 const authRoutes = require('./auth/routes');
-// const playerRouter = require('./routes/playerRoute.js');
-
+const gameRoutes = require('./routes/gameRoutes.js');
 const notFoundHandler = require('./error-handlers/404');
 const errorHandler = require('./error-handlers/500');
 
- 
 const app = express();
 
-// app.get('/hello',(req,res)=>{
-//   res.send('welcooooooome')
-// })
+
 app.use(cors());
 app.use(morgan('dev'));
 
@@ -24,8 +20,7 @@ app.use(express.urlencoded({ extended: true }));
 
 app.use(logger);
 app.use(authRoutes);
-
-// app.use('/api/v1/player/', playerRouter);
+app.use(gameRoutes);
 
 app.use('*', notFoundHandler);
 app.use(errorHandler);
